requirements.txt
Flask
pandas
SQLAlchemy
pyodbc
python-dotenv
yfinance